Fix NoneType errors when reading from a path.

+++ remote_offsetfinder.py
@@ -12,25 +12,29 @@ import utils
 def getDecryptionCMD(device: str, version: str) -> str:
     keys_path = api.getKeysForVersion(device, version)
     data = utils.readJSONFile(keys_path)
-    for key in data['keys']:
-        if key['image'] == 'Kernelcache':
-            cmd = (
-                '/usr/local/bin/xpwntool',
-                'OF32/kernelcache.encrypted',
-                'OF32/kernelcache.decrypted',
-                f'-iv {key["iv"]}',
-                f'-k {key["key"]}'
-            )
-            return ' '.join(cmd)
+    if data:
+        for key in data['keys']:
+            if key['image'] == 'Kernelcache':
+                cmd = (
+                    '/usr/local/bin/xpwntool',
+                    'OF32/kernelcache.encrypted',
+                    'OF32/kernelcache.decrypted',
+                    f'-iv {key["iv"]}',
+                    f'-k {key["key"]}'
+                )
+                return ' '.join(cmd)
 
 
-def downloadKernel(device: str, version: str) -> None:
+def downloadKernel(device: str, version: str) -> Path:
     device_path = api.getDeviceData(device)
     data = utils.readJSONFile(device_path)
-    url = api.getVersionURL(version, data)
-    kernel_path = Path(f'kernels/{device}/{version}')
-    kernel_path.mkdir(exist_ok=True)
-    api.downloadKernelFromURL(url, f'{kernel_path.name}/kernelcache.encrypted')
+    if data:
+        url = api.getVersionURL(version, data)
+        kernel_path = Path(f'kernels/{device}/{version}')
+        kernel_path.mkdir(parents=True, exist_ok=True)
+        kernel_encrypted = Path(f'{kernel_path.resolve()}/kernelcache.encrypted')
+        api.downloadKernelFromURL(url, kernel_encrypted.resolve())
+        return kernel_encrypted
 
 
 def getOF32CMD() -> str:
@@ -71,10 +75,10 @@ class Client:
         self.sftp.unlink(path)
 
     def uploadFile(self, file: Path, path: str) -> None:
-        self.sftp.put(file.name, path)
+        self.sftp.put(file.resolve(), path)
 
     def downloadFile(self, file: str, path: Path) -> None:
-        self.sftp.get(file, path.name)
+        self.sftp.get(file, path.resolve())
 
     def removeKernels(self) -> None:
         contents = self.listDir('OF32')
@@ -119,35 +123,35 @@ def prepareHomeDepotJSON(device: str, version: str, data: dict) -> dict:
 
 def getOffsets(address: str, user: str, password: str, device: str, version: str) -> None:
     print(f'[*] Downloading kernel for {device} {version}')
-    downloadKernel(device, version)
+    kernel_path = Path(f'kernels/{device}/{version}')
+    kernel_encrypted = downloadKernel(device, version)
     client = Client(address, user, password)
     client.removeKernels()
     print('[*] Uploading kernel')
-    kernel_path = Path(f'kernels/{device}/{version}')
-    kernel_encrypted = Path(f'{kernel_path.name}/kernelcache.encrypted')
-    client.uploadFile(kernel_encrypted, 'OF32/kernelcache.encrypted')
+    client.uploadFile(kernel_encrypted.resolve(), 'OF32/kernelcache.encrypted')
     print('[*] Decrypting kernel')
     client.runCMD(getDecryptionCMD(device, version))
     print('[*] Running OF32')
     client.runCMD(getOF32CMD())
-    client.downloadFile('OF32/kernelcache.encrypted', Path(f'{kernel_path}/kernelcache.decrypted'))
+    client.downloadFile('OF32/kernelcache.encrypted', Path(f'{kernel_path.resolve()}/kernelcache.decrypted'))
     client.removeKernels()
     print('[*] Reading offsets')
     offsets_raw = client.readFile('OF32/offsets.txt')
-    client.removeFile('OF32/offsets.txt')
+    if offsets_raw:
+        client.removeFile('OF32/offsets.txt')
+        print('[*] Parsing offsets')
+        parsed_offests = parseOffsets(offsets_raw)
+        if parsed_offests:
+            print('[*] Writing offsets to json files')
+            utils.updateJSONFile(Path('payload/offsets.json'), parsed_offests)
+            depot_offsets = prepareHomeDepotJSON(device, version, parsed_offests)
+            utils.updateJSONFile(Path('HomeDepot.json'), depot_offsets)
+        print('[*] Adding kernels to ZPAQ archive')
+        utils.appendFileToZPAQArchive(Path('kernels'), Path('kernels.zpaq'), 1)
     #################################
     client.sftp.close()  # IMPORTANT
     client.ssh.close()  # IMPORTANT
     #################################
-    print('[*] Parsing offsets')
-    parsed_offests = parseOffsets(offsets_raw)
-    if parsed_offests:
-        print('[*] Writing offsets to json files')
-        utils.updateJSONFile(Path('payload/offsets.json'), parsed_offests)
-        depot_offsets = prepareHomeDepotJSON(device, version, parsed_offests)
-        utils.updateJSONFile(Path('HomeDepot.json'), depot_offsets)
-    print('[*] Adding kernels to ZPAQ archive')
-    utils.appendFileToZPAQArchive(Path('kernels'), Path('kernels.zpaq'), 1)
     print('#'*100)
 
 
@@ -160,8 +164,9 @@ def getAllOffsetsForDevice(address: str, user: str, password: str, device: str) 
 def getAllOffsets(address: str, user: str, password: str) -> None:
     devices_path = api.getAllDevices()
     data = utils.readJSONFile(devices_path)
-    for device in data:
-        getAllOffsetsForDevice(address, user, password, device['identifier'])
+    if data:
+        for device in data:
+            getAllOffsetsForDevice(address, user, password, device['identifier'])
 
 
 def main(args: list) -> None:
